test: avoid poll and loadkeys test duplication
