+++ tests/e2e/Autoservicio/pos.autoservicio.base.e2e.spec.js
@@ -1,0 +1,11 @@
+const { setupSuiteContext } = require("../helpers/suiteContext");
+const { describePosBaseTests } = require("../shared/posBase.shared");
+
+describe("POS Autoservicio - POSBase", () => {
+    const suite = setupSuiteContext();
+
+    describePosBaseTests({
+        productName: "POS Autoservicio",
+        createPos: () => suite.createAutoservicio()
+    });
+});
