--- tests/e2e/Autoservicio/pos.autoservicio.poll.e2e.spec.js
@@ -1,35 +1,0 @@
-const { setupSuiteContext } = require("../helpers/suiteContext");
-const {
-    ACK_BYTE,
-    buildMessage,
-    captureSend,
-    sendReply,
-    connectWithPollAck
-} = require("../helpers/mockPos");
-
-const createConnectedPos = async (suite) => {
-    const pos = suite.createAutoservicio();
-    await connectWithPollAck(pos);
-    return pos;
-}
-
-describe("POS Autoservicio - Poll", () => {
-    const suite = setupSuiteContext();
-
-    it("checks connection with successful result", async () => {
-        const pos = await createConnectedPos(suite);
-        
-        const pollPromise = pos.poll();
-        const sentMessage = await captureSend(pos);
-        await sendReply(pos, ACK_BYTE);
-        expect(sentMessage).toEqual(buildMessage("0100"));
-        const response = await pollPromise;
-        expect(response).toBe(true);
-    });
-
-    it("checks connection with error result", async () => {
-        const pos = suite.createAutoservicio();
-
-        await expect(pos.poll()).rejects.toThrow("You have to connect to a POS");
-    });
-});
