+++ tests/e2e/Integrado/pos.integrado.base.e2e.spec.js
@@ -1,0 +1,11 @@
+const { setupSuiteContext } = require("../helpers/suiteContext");
+const { describePosBaseTests } = require("../shared/posBase.shared");
+
+describe("POS Integrado - POSBase", () => {
+    const suite = setupSuiteContext();
+
+    describePosBaseTests({
+        productName: "POS Integrado",
+        createPos: () => suite.createIntegrado()
+    });
+});
